Add support for AnimatedTransforms
These work by introducing a second transformation
matrix, so now there’s one at the start of the
frame and one at the end of the frame.

When the renderer sends rays out into the scene,
they have a time associated with them and the ray
may or may not hit an object depending on where it
is at that time - which is determined by
interpolating over the start/end transformations.