Add a method to write the scene description to a file
The pbrt gem supports stream to an IO object, so
rather than keep scene descriptions in memory it’s
probably better to write them to disk.

The tempfile is cleaned up when the block
finishes, so if I want to keep the description
file around I’ll need to copy it before then.