Remove the scene directory before rendering
If the scene has already been rendered, remove the
previous directory before rendering again. This
prevents stale frames being added onto the end of
the scene, if the previous scene had more frames
than this one.

If the scene is exactly the same as it was before,
the frames will be read from the cache, so we
won’t pay the price of rendering those frames
again.